apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-dotenv
yfinance
pandas
httpx
